Codechange: Move TimerGameCalendar functions to a parent TimerGameCommon class
